+++ knights_tour.h
@@ -14,10 +14,10 @@ class KnightsTour {
 public:
 	KnightsTour(int board_size);
 
-	int generate(int row, int col);
+	unsigned long long int generate(int row, int col);
 
 private:
-	void move(int row, int col, int& m, int& num_tours);
+	void move(int row, int col, int& m, unsigned long long int& num_tours);
   	void get_moves(int row, int col, int row_moves[], int col_moves[], int& num_moves);
 	void print();
